Reduce number of opening and closing a file.
To not opening and closing a file four times, we read
the file's contents, do a search and replace in memory,
seek to position 0 of the file and directly write the
changes.

Therefore, opening and closing a file is only necessary
two times instead of four times.

Additionally an editor can be defined via environment
variable EDITOR_NOTPAD.

Windows 10 fix found: truncate( FileHandle, 0 ) before writing.

Updated documentation that Windows 10 problem is fixed.